% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_filter_fns.R
\name{algApplyFn}
\alias{algApplyFn}
\title{Apply one of the Multiple Filter ensemble algorithms - enabling parallel computation in make_ensemble_parallle() function}
\usage{
algApplyFn(algorithmL, x, y)
}
\arguments{
\item{algorithmL}{The string corresponding to available caret package algorithm names for caret trainControl function.}

\item{x}{The dataset to train the algorithm.}

\item{y}{The binary variable output.}
}
\value{
Vector indicating whether a sample was predicted discordant by this algorithm.
}
\description{
Apply one of the Multiple Filter ensemble algorithms - enabling parallel computation in make_ensemble_parallle() function
}
\examples{
algApplyFn(algorithmL, x, y)
}
