% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_fns.R
\name{inject_label_noiseR}
\alias{inject_label_noiseR}
\title{Probabilistically inject label noise according to proportions}
\usage{
inject_label_noiseR(y, flip_i, flip_j)
}
\arguments{
\item{y}{The binary variable output.}

\item{flip_i}{The proportion of samples of class A to flip to class B - will probabilistically flip ~this proportion of samples, but may not be exact.}

\item{flip_j}{The proportion of samples of class B to flip to class A - will probabilistically flip ~this proportion of samples, but may not be exact.}
}
\value{
A list containing: "yz", the flipped sample labels; "fd", an indicator variable with +1 if the sample was flipped and -1 if the sample label is unflipped.
}
\description{
Probabilistically inject label noise according to proportions
}
