% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_fns.R
\name{intro_to_unittest}
\alias{intro_to_unittest}
\title{Test out how to write appropriate unit tests}
\usage{
intro_to_unittest(x)
}
\arguments{
\item{x}{integer value}
}
\value{
Text indicating whether the input was a "1"
}
\description{
Test out how to write appropriate unit tests
}
