% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_filter_fns.R
\name{make_ensemble_parallel}
\alias{make_ensemble_parallel}
\title{Multiple Filter ensemble algorithm, parallel version}
\usage{
make_ensemble_parallel(x, y, library)
}
\arguments{
\item{x}{The dataset to train the algorithm.}

\item{y}{The binary variable output.}

\item{library}{List of algorithm ID strings(corresponding to available caret package algorithm names for caret trainControl function) to use in the multiple filter ensemble.}
}
\value{
A list containing: "MF", samples to be kept by majority filter; "CF", samples to be kept by consensus filter, "full_res", the full list of discordant predictions
}
\description{
Multiple Filter ensemble algorithm, parallel version
}
\examples{
make_ensemble_parallel(x, y, c("rf","regLogistic","nnet"))
}
