% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_setup_fns.R
\name{split_train_test_cv}
\alias{split_train_test_cv}
\title{Given a dataset, split it into training and test sets randomly.}
\usage{
split_train_test_cv(input, regex, source_variable, flip_i, flip_j,
  cv = 10)
}
\arguments{
\item{input}{Dataset}

\item{regex}{The regular expression for this class.}

\item{source_variable}{The GEO dataset variable that will be used to split the dataset.}

\item{flip_i}{Decimal value indicating the proportion of samples of class B to flip to class A.}

\item{flip_j}{Decimal value indicating the proportion of samples of class B to flip to class A.}

\item{cv}{The number of cross-validation splits to return; default = 10.}
}
\value{
List of datasets split for cross-validation with training and test sets, including flipped labels at specified percentage.
}
\description{
Given a dataset, split it into training and test sets randomly.
}
