% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_setup_fns.R
\name{subset_geo_cv}
\alias{subset_geo_cv}
\title{Create a cross-validation dataset from a known GEO dataset with a covariate of interest.}
\usage{
subset_geo_cv(geo_dataset_name, geo_dataset_list, pos_regex, neg_regex,
  source_variable, flip_i, flip_j)
}
\arguments{
\item{geo_dataset_name}{The dataset name set initially.}

\item{geo_dataset_list}{The pre-loaded list of GEO datasets.}

\item{pos_regex}{A regular expression (string) that selects for the positive class in the GEO dataset metadta field specified by source_variable arg}

\item{neg_regex}{A regular expression (string) that selects for the negative class in the GEO dataset metadta field specified by source_variable arg}

\item{source_variable}{The GEO dataset variable that will be used to split the dataset.}

\item{flip_i}{Decimal value indicating the proportion of samples of class B to flip to class A.}

\item{flip_j}{Decimal value indicating the proportion of samples of class B to flip to class A.}
}
\value{
List of datasets split for cross-validation with training and test sets, including flipped labels at specified percentage.
}
\description{
Create a cross-validation dataset from a known GEO dataset with a covariate of interest.
}
\examples{
subset_geo_cv(geo_dataset_name, geo_dataset_list, source_variable, flip_i, flip_j)
}
