% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_fns.R
\name{castLabel}
\alias{castLabel}
\title{Cast the labels to the appropriate notation for processing}
\usage{
castLabel(y, t)
}
\arguments{
\item{y}{The binary variable output.}

\item{t}{integer value indicating how to cast label values}
}
\value{
y, a list of labels cast to the appropriate notation
}
\description{
Cast the labels to the appropriate notation for processing
}
